Migrated handleMasterMsg, moved some functions to the right modules
